Work on optimising interleaving.

+++ Rendering/vtkShellExtractor.h
@@ -1,7 +1,7 @@
 // vtkShellExtractor.h copyright (c) 2003 
 // by Charl P. Botha cpbotha@ieee.org 
 // and the TU Delft Visualisation Group http://visualisation.tudelft.nl/
-// $Id: vtkShellExtractor.h,v 1.6 2004/01/15 11:00:55 cpbotha Exp $
+// $Id: vtkShellExtractor.h,v 1.7 2004/06/16 16:13:42 cpbotha Exp $
 // vtk class for extracting Udupa Shells
 
 /*
@@ -38,7 +38,7 @@ struct ShellVoxel
     /**
      * x-coordinate of voxel.
      */
-    int x;
+    unsigned short x;
     //float Value;
     float Opacity;
     float Red;
